v1.5.0: Architecture split implementation
✅ NEW ARCHITECTURE:
- app/overtake.py: Convergence zone analysis for 18 segments with overtake_flag='y'
- app/density.py: Universal density analysis for all 36 segments
- app/report.py: Combined reporting functionality
- app/main.py: Updated API endpoints with new architecture

✅ SAFETY BACKUPS:
- app/density_old.py: Preserves working density calculations
- app/main_old.py: Preserves existing API structure
- data/overlaps.csv: Kept for reference

✅ VALIDATION RESULTS:
- A1c Segment: Convergence at 1.8km, 35 10K overtaking 684 Half runners
- B1 Segment: Convergence at 3.54km, 67 10K overtaking 29 Full runners
- Total: 36 density segments + 18 overtake segments = 4 with convergence
- All modules import and execute successfully

✅ NEW ENDPOINTS:
- /api/density: Universal density analysis
- /api/overtake: Selective overtake analysis
- /api/report: Combined reporting
- /api/overlap: Legacy compatibility

🎯 READY FOR TESTING - NOT MERGED TO MAIN YET
